Refactor Types
